Deploy contract to evm properly.
When deploy a contract bytecode, we should create a new address
first and use it as callee account when invoking evm with deploy
bytecode, so that variable initialization could be stored in
the correct namespace (address of new contract), instead of zero
address, as what we did prior to this commit.

Also, let's use encoded string as keys when get/put state, instead
of string(raw bytes), to improve human-readability in logs.

Change-Id: I146cb0be6eabfb368798ea7f2ecceb9ab7feb002